change StartupProbe to ReadinessProbe
In some cases, the launcher container may start slowly, and using StartupProbe here may cause the container to restart and the entire task to fail.
In addition, the criteria for health check has been modified to ensure that the health check is declared successful only after the launcher is truly ready.