Removed absolute path to hydra_pmi_proxy in cmdArgs of worker pods to improve compatibility
In some images, the hydra_pmi_proxy is placed in /usr/local/bin/, while in others, it is placed in /usr/bin/, as indicated in the log of the launcher task - "Launch arguments: /usr/bin/hydra_pmi_proxy --control-port wrf-2-875ef-launcher-6wtjf:20000 ...". In this commit, the absolute path has been removed to ensure compatibility.